-Fixing platform paths for pjsip and risip for each seperate platform. -Several UI fixes and call dialog has now more proper status and call details. -Android builds are easier now. -RisipBuddy works better with numbers -RisipCallManager adjustments to work better with SIP and phone number calls.

+++ src/risip/models/risipcallhistorymodel.cpp
@@ -83,7 +83,7 @@ QVariant RisipCallHistoryModel::data(const QModelIndex &index, int role) const
     case CallDirectionRole:
         return call->callDirection();
     case CallDurationRole:
-        return call->callDuration();
+        return (qlonglong)call->callDuration();
     case CallTimestampRole:
         return call->timestamp();
     case CallContactRole:
